feat: Enhance logging and introduce new architecture documentation
This commit updates the logging functionality in the application by adding timestamps and log levels to messages, improving the clarity and traceability of logs. Additionally, a comprehensive architecture document has been introduced, detailing the structure and organization of the Tauri backend for the POS application, including directory layouts, architecture layers, and key design principles. This enhances maintainability and provides a clear reference for future development.
