+++ desktop/src-tauri/src/commands/mod.rs
@@ -1,0 +1,12 @@
+pub mod auth;
+pub mod dte;
+pub mod secure_storage;
+pub mod system;
+pub mod utils;
+
+// Re-export all commands for easy access
+pub use auth::*;
+pub use dte::*;
+pub use secure_storage::*;
+pub use system::*;
+// utils module is currently empty, so no need to re-export
