+++ desktop/src-tauri/src/commands/utils.rs
@@ -1,0 +1,8 @@
+// Utility commands module
+// This module is reserved for general utility commands that don't fit into specific categories
+
+// Example placeholder for future utility commands:
+// #[tauri::command]
+// pub fn some_utility_function() -> Result<String, String> {
+//     Ok("Utility function result".to_string())
+// }
