+++ desktop/src-tauri/src/services/mod.rs
@@ -1,0 +1,3 @@
+pub mod database;
+pub mod dte_signer;
+pub mod secure_storage;
